♻️ refactor: DTO 구조 변경
