--- user-service/src/main/java/com/waither/userservice/dto/EmailVerificationDto.java
@@ -1,9 +1,0 @@
-package com.waither.userservice.dto;
-
-public record EmailVerificationDto(
-        String email,
-        String authCode
-) {
-
-}
-
