+++ user-service/src/main/java/com/waither/userservice/dto/response/SettingResDto.java
@@ -1,0 +1,4 @@
+package com.waither.userservice.dto.response;
+
+public class SettingResDto {
+}
