+++ user-service/src/main/java/com/waither/userservice/dto/response/UserResDto.java
@@ -1,0 +1,5 @@
+package com.waither.userservice.dto.response;
+
+public class UserResDto {
+
+}
